Update where edits save based on login
Changed the search page, so that if you try and navigate to it when
you are not logged in, it will redirect you to the volunteerSearch
page. Also updated the edit page, so it registers when you are
logged in as an admin and saves your edits directly to the proper
collections (rather than putting them in the proposed collection).